% Generated by roxygen2 (4.0.2): do not edit by hand
\name{correlationAnalysis}
\alias{correlationAnalysis}
\title{correlationAnalysis}
\usage{
correlationAnalysis(expression, pairs2Target = pairs2TargetRemDup, n = 300)
}
\arguments{
\item{expression}{the expression matrix, can be the result of \code{\link{prepareGeoData}}.}

\item{pairs2Target}{the targets of tf pairs}

\item{n}{the number of genes selected for prior comparison, to make the analysis faster}
}
\description{
A function to perform correlation analysis for expression data, to see if the shared targets of two tf have significantly higher correlations.
}
\details{
A function to perform correlation analysis for expression data, to see if the shared targets of two tf have significantly higher correlations.
}
\examples{
1
}

