% Generated by roxygen2 (4.0.2): do not edit by hand
\name{differentialAnalysis}
\alias{differentialAnalysis}
\title{differentialAnalysis}
\usage{
differentialAnalysis(x, groups, contrasts, LogC = NULL,
  pairs2Target = pairs2TargetRemDup, pCut = 0.05, fCut = log2(1.5))
}
\arguments{
\item{x}{the expression data matrix, can be result from \code{\link{prepareGeoData}}.}

\item{groups}{indicate the samples belong to which group.}

\item{contrasts}{indicate groups to be compared.}

\item{LogC}{logical. Indicate if a log2 transformation will be performed. If not specified, the function will try to detect it.}

\item{pairs2Target}{the targets of tf pairs}

\item{pCut}{the cutoff of p value for significant genes}

\item{fCut}{the cutoff of fold change for significant genes}
}
\description{
A wrapper for function performLimma and enrichTarget.
}
\details{
A function to perform differential analysis for expression data, to see if the shared targets of two tf were significantly changed between two conditions.
}
\examples{
\dontrun{
dataMatrix<-prepareGeoData(GEO="GDS2213")
#use TF pairs in HEPG2 cell only
tfPairsUsed<-pairs2TargetRemDup[which(pairs2TargetRemDup[,3]=="HEPG2"),]
enrichTargetResult<-differentialAnalysis(dataMatrix,groups=c(rep("untreated",4),rep("aza",4),rep("TSA",4),rep("azaAndTSA",4)),contrasts=c("aza-untreated","TSA-untreated","azaAndTSA-untreated"),pairs2Target=tfPairsUsed)
enrichTargetResultNetwork1<-networkVis(enrichTargetResult[[1]])
enrichTargetResultNetwork3<-networkVis(enrichTargetResult[[3]])
}
}

