% Generated by roxygen2 (4.0.2): do not edit by hand
\name{enrichTarget}
\alias{enrichTarget}
\title{enrichTarget}
\usage{
enrichTarget(limmaResult, pairs2Target = pairs2TargetRemDup, pCut = 0.05,
  fCut = log2(1.5))
}
\arguments{
\item{limmaResult}{the result of \code{\link{performLimma}}.}

\item{pairs2Target}{the targets of tf pairs}

\item{pCut}{the cutoff of p value for significant genes}

\item{fCut}{the cutoff of fold change for significant genes}
}
\description{
A function to perform enrichment analysis in limma result, to see if the shared targets of two tf were enriched in significant different genes.
}
\details{
A function to perform enrichment analysis in limma result, to see if the shared targets of two tf were enriched in significant different genes.
}
\examples{
\dontrun{
dataMatrix<-prepareGeoData(GEO="GDS2213")
limmaResult<-performLimma(dataMatrix,groups=c(rep("untreated",4),rep("aza",4),rep("TSA",4),rep("azaAndTSA",4)),contrasts=c("aza-untreated","TSA-untreated","azaAndTSA-untreated"))
tfPairsUsed<-pairs2TargetRemDup[which(pairs2TargetRemDup[,3]=="HEPG2"),]
enrichTargetResult<-enrichTarget(limmaResult,pairs2Target=tfPairsUsed)
}
}

