% Generated by roxygen2 (4.0.2): do not edit by hand
\name{networkVis}
\alias{networkVis}
\title{networkVis}
\usage{
networkVis(result, pCut = 0.05)
}
\arguments{
\item{result}{the expression matrix, can be the result of \code{\link{prepareGeoData}}.}

\item{pCut}{the p value cutoff for genes used in network visualization}
}
\description{
A function to perform correlation analysis for expression data, to see if the shared targets of two tf have significantly higher correlations.
}
\details{
A function to perform correlation analysis for expression data, to see if the shared targets of two tf have significantly higher correlations.
}
\examples{
\dontrun{
dataMatrix<-prepareGeoData(GEO="GDS2213")
#use TF pairs in HEPG2 cell only
tfPairsUsed<-pairs2TargetRemDup[which(pairs2TargetRemDup[,3]=="HEPG2"),]
enrichTargetResult<-differentialAnalysis(dataMatrix,groups=c(rep("untreated",4),rep("aza",4),rep("TSA",4),rep("azaAndTSA",4)),contrasts=c("aza-untreated","TSA-untreated","azaAndTSA-untreated"),pairs2Target=tfPairsUsed)
#Show the network for contrast 1
enrichTargetResultNetwork1<-networkVis(enrichTargetResult[[1]])
#Show the network for contrast 3, which is different from contrast 1
enrichTargetResultNetwork3<-networkVis(enrichTargetResult[[3]])
}
}

