% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{pairs2TargetRemDup}
\alias{pairs2TargetRemDup}
\title{transcription factors pairs and their targets from ENCODE data.}
\source{
Gerstein MB, et al. (2012) Architecture of the human regulatory
network derived from ENCODE data, Nature, Sep 6;489(7414):91-100.
}
\usage{
pairs2TargetRemDup
}
\description{
This data set is a data.frame with 9 columns. The first 6 columns were from the sTable 3 of reference. The last 3 columns were targets for TF1 and TF2, TF1 only, TF2 only.
}

