% Generated by roxygen2 (4.0.2): do not edit by hand
\name{performLimma}
\alias{performLimma}
\title{performLimma}
\usage{
performLimma(x, groups, contrasts, LogC)
}
\arguments{
\item{x}{the expression data matrix, can be result from \code{\link{prepareGeoData}}.}

\item{groups}{indicate the samples belong to which group.}

\item{contrasts}{indicate groups to be compared.}

\item{LogC}{logical. Indicate if a log2 transformation will be performed. If not specified, the function will try to detect it.}
}
\description{
A function to perform limma analysis in prepared geo data.
}
\details{
A function to perform limma analysis in prepared geo data.
}
\examples{
\dontrun{
dataMatrix<-prepareGeoData(GEO="GDS2213")
limmaResult<-performLimma(dataMatrix,groups=c(rep("untreated",4),rep("aza",4),rep("TSA",4),rep("azaAndTSA",4)),contrasts=c("aza-untreated","TSA-untreated","azaAndTSA-untreated"))
}
}

