% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prepareGeoData}
\alias{prepareGeoData}
\title{prepareGeoData}
\usage{
prepareGeoData(GEO = NULL, destdir = getwd(),
  geneSymColumnName = "Gene Symbol")
}
\arguments{
\item{GEO}{the GEO ID.}

\item{destdir}{indicate the path to download the files.}

\item{geneSymColumnName}{indicate the column name for gene symbol, which will be used in ID converter For GSE data.}
}
\description{
A function to download GDS or GSE data and convert ID if needed.
}
\details{
A function to download GDS or GSE data and convert ID if needed.
}
\examples{
dataMatrix<-prepareGeoData("GDS2213")
}

